requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.9
gunicorn==23.0.0
inflection==0.5.1
mysqlclient==2.2.6
packaging==24.2
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
whitenoise==6.9.0
